GUACAMOLE-377: Use FreeRDP's built-in GDI handling of bitmap cache.
With FreeRDP's GDI being used for ther other aspects of RDP drawing,
using our own implementation of bitmap caching causes the FreeRDP's GDI
to become out-of-sync with Guacamole's representation, producing
graphical artifacts. We can't simply monkey-patch the GDI, as the
functions used are internal and not part of the public FreeRDP API.

There are likely other possible approaches, like manually updating
FreeRDP's GDI surface in addition to Guacamole's surface, but this may
not be worth the effort given that bitmap caching is not commonly used
by modern RDP servers.